838 Rename Form to CustomForm::Form and Question to CustomForm::Question (#864)
* Added module custom_form for Form and Question Controller and Views

* Fix test cases for custom_form module for Form and Question

* Nested modules

* Modularize Form model and its related test cases

* Modularize Question Module with CustomForm

* Lint Fix

* Fix seed file

* Modulaize Question policy code and fix link in question form
