+++ app/policies/organizations/custom_form/form_policy.rb
@@ -1,0 +1,14 @@
+module Organizations
+  module CustomForm
+    class FormPolicy < ApplicationPolicy
+      pre_check :verify_organization!
+      pre_check :verify_active_staff!
+
+      alias_rule :new?, :create?, :index?, to: :manage?
+
+      def manage?
+        permission?(:manage_forms)
+      end
+    end
+  end
+end
