+++ app/policies/organizations/custom_form/question_policy.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Organizations
+  module CustomForm
+    class QuestionPolicy < ApplicationPolicy
+      pre_check :verify_organization!
+      pre_check :verify_active_staff!
+
+      alias_rule :new?, :create?, :index?, to: :manage?
+
+      def manage?
+        permission?(:manage_questions)
+      end
+    end
+  end
+end
