+++ app/views/layouts/dashboard/_sidebar.html.erb
@@ -61,7 +61,7 @@
       </li>
       <% if current_user.staff_account && current_user.has_role?(:admin, current_user.staff_account.organization) %>
         <li class="nav-item">
-          <%= active_link_to staff_forms_path, class: 'nav-link' do %>
+          <%= active_link_to staff_custom_form_forms_path, class: 'nav-link' do %>
             <i class="nav-icon fe fe-clipboard me-2"></i>Forms
           <% end %>
         </li>
