+++ app/views/organizations/staff/custom_form/forms/_form.html.erb
@@ -18,7 +18,7 @@
       </div>
 
       <%= f.submit t('general.save'), class: 'btn btn-primary' %>
-      <%= link_to t('general.cancel'), staff_forms_path, class: 'btn btn-secondary' %>
+      <%= link_to t('general.cancel'), staff_custom_form_forms_path, class: 'btn btn-secondary' %>
     <% end %>
   </div>
 </div>
