+++ config/locales/en.yml
@@ -539,57 +539,58 @@ en:
               paused: "Paused"
               published: "Published"
               draft: "Draft"
-      forms:
-        set_form: 
+      custom_form:
+        forms:
+          set_form: 
+            not_found: "Form not found."
           not_found: "Form not found."
-        not_found: "Form not found."
-        index:
-          forms: "Forms"
-          create_form: "Create Form"
-          confirm_delete: "Are you sure you want to delete this form?"
-        new:
-          new_form: "New Form"
-        create:
-          success: "Form saved successfully."
-          error: "Error creating form."
-        edit:
-          edit_form: "Edit Form"
-        show:
-          confirm_delete_question: "Are you sure you want to delete this question?"
-          create_question: "Create Question"
-        update:
-          success: "Form saved successfully."
-        destroy:
-          success: "Form was successfully deleted."
-        form:
-          help_text:
-            title: "What the user sees as the title of the form."
-            instructions: "Additional instructions to aid the user in filling out the form."
-        saved: "Form saved successfully."
-        updated: "Form saved successfully."
-        destroyed: "Form was successfully deleted."
-      questions:
-        set_question:
-          not_found: "Not Found"
-        new:
-          new_question: "New Question"
-        create:
-          success: "Question saved successfully."
-          error: "Error creating question."
-        update:
-          success: "Question saved successfully."
-          error: "Error updating question."
-        updated: "Question saved successfully."
-        saved: "Question saved successfully translated."
-        destroy:
-          success: "Question was successfully deleted."
-        not_found: "Question not found."
-        edit:
-          edit_question: "Edit Question"
-        form:
-          help_text:
-            label: "Label"
-            help_text: "help text"
+          index:
+            forms: "Forms"
+            create_form: "Create Form"
+            confirm_delete: "Are you sure you want to delete this form?"
+          new:
+            new_form: "New Form"
+          create:
+            success: "Form saved successfully."
+            error: "Error creating form."
+          edit:
+            edit_form: "Edit Form"
+          show:
+            confirm_delete_question: "Are you sure you want to delete this question?"
+            create_question: "Create Question"
+          update:
+            success: "Form saved successfully."
+          destroy:
+            success: "Form was successfully deleted."
+          form:
+            help_text:
+              title: "What the user sees as the title of the form."
+              instructions: "Additional instructions to aid the user in filling out the form."
+          saved: "Form saved successfully."
+          updated: "Form saved successfully."
+          destroyed: "Form was successfully deleted."
+        questions:
+          set_question:
+            not_found: "Not Found"
+          new:
+            new_question: "New Question"
+          create:
+            success: "Question saved successfully."
+            error: "Error creating question."
+          update:
+            success: "Question saved successfully."
+            error: "Error updating question."
+          updated: "Question saved successfully."
+          saved: "Question saved successfully translated."
+          destroy:
+            success: "Question was successfully deleted."
+          not_found: "Question not found."
+          edit:
+            edit_question: "Edit Question"
+          form:
+            help_text:
+              label: "Label"
+              help_text: "help text"
       staff:
         deactivate:
           success: "Staff account deactivated."
