+++ config/routes.rb
@@ -46,8 +46,10 @@ Rails.application.routes.draw do
       resources :staff_invitations, only: %i[new]
       resources :fosterer_invitations, only: %i[new]
 
-      resources :forms do
-        resources :questions
+      namespace :custom_form do
+        resources :forms do
+          resources :questions
+        end
       end
       post "user_roles/:id/to_staff", to: "user_roles#to_staff", as: "user_to_staff"
       post "user_roles/:id/to_admin", to: "user_roles#to_admin", as: "user_to_admin"
